Refactor code analysis code.
